Make representation handler classes package-private and move internal package components up to ReprUtil
