+++ src/main/java/org/cryptimeleon/math/serialization/annotations/RepresentationHandler.java
@@ -1,4 +1,4 @@
-package org.cryptimeleon.math.serialization.annotations.internal;
+package org.cryptimeleon.math.serialization.annotations;
 
 import org.cryptimeleon.math.serialization.Representation;
 import org.cryptimeleon.math.serialization.annotations.RepresentationRestorer;
@@ -8,7 +8,7 @@ import java.util.function.Function;
 /**
  * Interface for classes that can serialize and deserialize specific types of objects.
  */
-public interface RepresentationHandler {
+interface RepresentationHandler {
     /**
      * Deserializes the given representation using the given representation restorers.
      * @param repr the representation to deserialize
