--- src/main/java/org/cryptimeleon/math/serialization/annotations/internal/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Implements handlers for converting Java objects to and from representations.
- */
-package org.cryptimeleon.math.serialization.annotations.internal;
